refactor: seed workspace domain package
